fix: исправлены модели UserModel, AdModel, CommentModel с аннотациями Hibernate и Swagger fix: улучшены DTO классы согласно OpenAPI-спецификации fix: изменены контроллеры UserController, AdController, CommentController, AuthController с аннотациями OpenAPI feat: добавлены сервисы UserServiceImpl, AdServiceImpl, CommentServiceImpl, AuthServiceImpl с базовой логикой и логированием fix: исправлен маппинг между сущностями и DTO через UserMapper, AdMapper, CommentMapper feat: добавлена авторизация и регистрация пользователя с использованием Spring Security feat: реализована логика смены пароля пользователем и обновления профиля feat: реализовано добавление, обновление, получение и удаление объявлений с загрузкой изображений feat: реализовано добавление, обновление, получение и удаление комментариев к объявлениям feat: реализовано получение объявлений и комментариев авторизованного пользователя feat: добавлены валидаторы и сообщения об ошибках для всех DTO согласно OpenAPI feat: добавлены глобальные исключения и собственные Exception-классы fix: исправлен маппинг ExtendedAd с учётом конструктора с полным набором полей fix: обновлён AdServiceImpl под реальные изображения и работу с репозиторием fix: реализована загрузка аватарки пользователя и сохранение в файловую систему fix: исправлена логика loginUser и registerUser для обработки ошибок и соответствия статусам fix: корректно возвращаются 400/401/403/404 коды ошибок по OpenAPI спецификации fix: добавлены тесты для всех контроллеров: AuthController, UserController, AdController, CommentController feat: добавлены custom Exceptions

+++ src/main/java/ru/skypro/homework/controller/AdController.java
@@ -53,7 +53,7 @@ public class AdController {
             @ApiResponse(responseCode = "404", description = "Объявление с таким ID не найдено")
     })
     @GetMapping("/{id}")
-    public ResponseEntity<ExtendedAd> getAd(@PathVariable Integer id) {
+    public ResponseEntity<ExtendedAd> getAd(@PathVariable Long id) {
         return ResponseEntity.ok(adService.getAdById(id));
     }
 
@@ -64,7 +64,7 @@ public class AdController {
             @ApiResponse(responseCode = "401", description = "Пользователь не авторизован")
     })
     @PatchMapping("/{id}")
-    public ResponseEntity<Ad> updateAd(@PathVariable Integer id, @RequestBody @Valid CreateOrUpdateAd ad) {
+    public ResponseEntity<Ad> updateAd(@PathVariable Long id, @RequestBody @Valid CreateOrUpdateAd ad) {
         return ResponseEntity.ok(adService.updateAd(id, ad));
     }
 
@@ -75,7 +75,7 @@ public class AdController {
             @ApiResponse(responseCode = "401", description = "Пользователь не авторизован")
     })
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteAd(@PathVariable Integer id) {
+    public ResponseEntity<Void> deleteAd(@PathVariable Long id) {
         adService.deleteAd(id);
         return ResponseEntity.noContent().build();
     }
@@ -97,7 +97,7 @@ public class AdController {
             @ApiResponse(responseCode = "404", description = "Объявление с таким ID не найдено"),
             @ApiResponse(responseCode = "401", description = "Пользователь не авторизован")
     })
-    public ResponseEntity<Void> updateAdImage(@PathVariable Integer id,
+    public ResponseEntity<Void> updateAdImage(@PathVariable Long id,
                                               @RequestPart("image") MultipartFile image) {
         adService.updateAdImage(id, image);
         return ResponseEntity.ok().build();
